# openapi.yaml
openapi: 3.0.3
info:
  title: Namimado Browser Control API
  version: 0.1.0
  description: |
    Headless control plane for the namimado desktop browser.

    **One spec, many faces.** This document is the single source of
    truth for every interface surface:

    - `namimado serve --port N` — HTTP REST server (axum)
    - `namimado mcp` — MCP server (rmcp stdio, via kaname)
    - Generated SDKs (via pleme-io `forge-gen`)
    - Generated client types (via pleme-io `takumi` / `sekkei`)

    The browser's Lisp substrate (12 def* DSLs: transform / scrape /
    alias / plan / predicate / agent / state / effect / derived /
    query / route / component) runs on every navigate; this API
    exposes both triggering it and observing its outcome.
  contact:
    name: pleme-io
    url: https://github.com/pleme-io/namimado
  license:
    name: MIT

servers:
  - url: http://localhost:7860
    description: Default local bind

tags:
  - name: control
    description: Drive the browser — navigate, reload, close.
  - name: observe
    description: Inspect substrate state — frameworks, effects, transforms.
  - name: meta
    description: Server metadata + the OpenAPI spec itself.

paths:
  /status:
    get:
      tags: [meta]
      summary: Service liveness + feature inventory.
      operationId: getStatus
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'

  /navigate:
    post:
      tags: [control]
      summary: Navigate to a URL and run the full Lisp substrate pipeline.
      description: |
        Fetches the URL, parses DOM + CSS, runs framework detection,
        state-bind, route-match, query dispatch, derived-aware effects,
        agent decisions, and transform+component apply. Returns the
        structured report observing every phase.
      operationId: navigate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NavigateRequest'
      responses:
        '200':
          description: Substrate report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavigateResponse'
        '400':
          description: Bad request (malformed URL, fetch failure, …)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /report:
    get:
      tags: [observe]
      summary: Structured substrate report from the last navigate.
      operationId: getLastReport
      responses:
        '200':
          description: Report from the most recent navigate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '404':
          description: No navigate has happened yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /state:
    get:
      tags: [observe]
      summary: Current substrate state-store snapshot.
      description: |
        Every `(defstate …)` cell's current value, accumulated across
        every navigate in this session. Same shape as the embedded
        `state_snapshot` field on `ReportResponse`.
      operationId: getState
      responses:
        '200':
          description: Array of state cells.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StateCellValue'

  /openapi.yaml:
    get:
      tags: [meta]
      summary: This document, YAML.
      operationId: getOpenapiYaml
      responses:
        '200':
          description: OK
          content:
            application/yaml:
              schema: { type: string }

  /openapi.json:
    get:
      tags: [meta]
      summary: This document, JSON.
      operationId: getOpenapiJson
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: object }

components:
  schemas:
    StatusResponse:
      type: object
      required: [service, version, features]
      properties:
        service: { type: string, example: namimado }
        version: { type: string, example: "0.1.0" }
        features:
          type: array
          items: { type: string, enum: [browser-core, gpu-chrome, http-server] }
        last_url: { type: string, nullable: true, format: uri }

    NavigateRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: Full URL or bare host. `example.com` → `https://example.com`.
          example: https://example.com

    NavigateResponse:
      type: object
      required: [final_url, fetched_bytes, report]
      properties:
        final_url: { type: string, format: uri }
        fetched_bytes: { type: integer, minimum: 0 }
        title: { type: string, nullable: true }
        report: { $ref: '#/components/schemas/ReportResponse' }

    ReportResponse:
      type: object
      required: [frameworks, queries_dispatched, effects_fired, agents_fired, transforms_applied, transform_hits, state_snapshot, derived_snapshot]
      properties:
        frameworks:
          type: array
          items: { $ref: '#/components/schemas/FrameworkHit' }
        routes_matched:
          type: string
          nullable: true
          description: Name of the `(defroute …)` that matched, if any.
        queries_dispatched:
          type: array
          items: { type: string }
          description: Names of `(defquery …)` forms fired via route on-match.
        effects_fired: { type: integer, minimum: 0 }
        agents_fired: { type: integer, minimum: 0 }
        transforms_applied: { type: integer, minimum: 0 }
        transform_hits:
          type: array
          items: { type: string }
          description: One entry per transform-applied-to-element.
        state_snapshot:
          type: array
          items: { $ref: '#/components/schemas/StateCellValue' }
        derived_snapshot:
          type: array
          items: { $ref: '#/components/schemas/StateCellValue' }

    FrameworkHit:
      type: object
      required: [name, confidence]
      properties:
        name: { type: string, example: NextJs }
        confidence: { type: number, format: float, minimum: 0, maximum: 1 }

    StateCellValue:
      type: object
      required: [name, value]
      properties:
        name: { type: string }
        value: { description: Arbitrary JSON — matches the state store shape. }

    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Machine-readable error code. }
        detail: { type: string, nullable: true, description: Human-readable context. }